+++ NEWS.md
@@ -1,4 +1,14 @@
-# manymome 0.3.3.2
+# manymome 0.3.3.4
+
+## Improvements
+
+- The `plot`-method of `cond_indirect_effects`
+  objects will keep the order of the
+  moderator levels (from bottom to top)
+  as appeared in the object.
+  Set `keep_wlevels_order` to `FALSE` to
+  revert to the old behavior.
+  (0.3.3.4)
 
 ## Miscellaneous
 
@@ -13,6 +23,14 @@
 
 - Updated `lavaan.mi` test objects
   for `lavaan 0.6-21`. (0.3.3.2)
+
+## Bug Fixes
+
+- Fixed the error message
+  of `plot.cond_indirect_effects()`
+  when a model has no latent
+  variables but `x_method` is set to
+  `"percentile"`. (0.3.3.3)
 
 # manymome 0.3.3